feat(rating): make personal rating fractional (1.0–10.0, step 0.1)
Personal rating moves from int to double. New FractionalStarRating widget
(tap/drag, leading clear cell, partial fill) replaces StarRatingBar inline
and in the table popup (live edit + OK / tap-outside to apply). Badges and
detail screen render one decimal. AniList import keeps precision (85 → 8.5).
.xcoll/.xcollx format bumps to v3; v2 files still import (legacy ints load
as doubles). DB column stays INTEGER via SQLite type affinity — no migration.

+++ lib/core/database/dao/collection_dao.dart
@@ -518,7 +518,7 @@ class CollectionDao {
     );
   }
 
-  Future<void> updateItemUserRating(int id, int? rating) async {
+  Future<void> updateItemUserRating(int id, double? rating) async {
     final Database db = await _getDatabase();
     await db.update(
       'collection_items',
